RET-1973. Update validations for date of birth. Cleaned the code around date validation.
